requirements.txt
# BRX Sync Microservice - Enterprise Grade
# Python 3.12+ Required

# FastAPI & ASGI Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database (PostgreSQL 16 + Async + MySQL)
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0
greenlet>=3.0.0
nest-asyncio>=1.6.0

# Task Queue
celery>=5.3.0
redis>=5.0.0
hiredis>=2.2.0

# Security
cryptography>=41.0.0
PyJWT>=2.8.0

# HTTP Client (for CardTrader API)
httpx>=0.25.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# UUID v7 Support
uuid6>=2023.5.2

# AWS Integration (optional, for SSM)
boto3>=1.29.0

# Metrics & Observability
prometheus-client>=0.19.0

# Optional: Development
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
